Revert "Migrate AuthTktCookieHelper to pyramid.security."
This reverts commit 9f267dd842c5e93336f0392f2809da75a716039a.
